test for controller added - organized controllers into their own controller packages
