+++ src/test/java/com/application/rezz/RezzApplicationTests.java
@@ -1,13 +1,21 @@
 package com.application.rezz;
 
+import com.application.rezz.controllers.LandingPageController;
 import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringBootTest
 class RezzApplicationTests {
 
+    @Autowired
+    private LandingPageController landingPageController;
+
     @Test
     void contextLoads() {
+        assertThat(landingPageController).isNotNull();
     }
 
 }
