+++ src/test/java/com/application/rezz/controllers/LandingPageControllerTests.java
@@ -1,0 +1,25 @@
+package com.application.rezz.controllers;
+
+
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.client.TestRestTemplate;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+public class LandingPageControllerTests {
+    @Value(value="${local.server.port}")
+    private int port;
+
+    @Autowired
+    private TestRestTemplate testRestTemplate;
+
+    @Test
+    public void testHiFunctionality() throws Exception {
+        assertThat(this.testRestTemplate.getForObject("http://localhost:" + port, String.class).contains("this"));
+    }
+
+}
